extended the authn and authz implementation to the infant controller.

+++ Controllers/InfantController.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 using BabyTracker.Models;
 using BabyTracker.Models.ViewModels;
@@ -15,18 +16,35 @@ namespace BabyTracker.Controllers
     public class InfantController: Controller
     {
         private BabyTrackerContext context;
-        public InfantController(BabyTrackerContext ctx)
+        private UserManager<IdentityUser> userManager;
+        private bool IsLoggedIn() => User.Identity.IsAuthenticated;
+        private bool IsInfantOwner(Infant infant) => infant.UserId == userManager.GetUserId(User);
+        public InfantController(BabyTrackerContext ctx, UserManager<IdentityUser> usrMgr)
         {
             context = ctx;
+            userManager = usrMgr;
         }
         public IActionResult Index()
         {
-            return View(context.Infants);
-        }
+            if (!IsLoggedIn())
+            {
+                return RedirectToPage("/Account/Login");
+            }
 
+            return View("Index", context.Infants.Where(i => i.UserId == userManager.GetUserId(User)));
+
+        }
         public async Task<IActionResult> Details (int id)
         {
+            if (!IsLoggedIn())
+            {
+                return RedirectToPage("/Account/Login");
+            }
             Infant i = await context.Infants.FirstOrDefaultAsync(i => i.InfantId == id);
+            if (!IsInfantOwner(i))
+            {
+                return RedirectToPage("/Error/Error404");
+            }
             InfantViewModel model = InfantViewModelFactory.Details(i);
             return View("InfantEditor", model);
         }
@@ -34,12 +52,27 @@ namespace BabyTracker.Controllers
         // HTTP GET
         public IActionResult Create()
         {
-            return View("InfantEditor", InfantViewModelFactory.Create(new Infant{Dob=DateTime.Now}));
+            if (!IsLoggedIn())
+            {
+                return RedirectToPage("/Account/Login");
+            }
+
+            Infant infant = new Infant
+            {
+                Dob = DateTime.Now,
+                UserId = userManager.GetUserId(User)
+            };
+
+            return View("InfantEditor", InfantViewModelFactory.Create(infant));
         }
 
         [HttpPost]
         public async Task<IActionResult> Create([FromForm] Infant infant)
         {
+            if (!IsLoggedIn())
+            {
+                return RedirectToPage("/Account/Login");
+            }
             if (ModelState.IsValid)
             {
                 infant.InfantId = default;
@@ -53,14 +86,31 @@ namespace BabyTracker.Controllers
         // HTTP GET
         public async Task<IActionResult> Edit (long id)
         {
+            if (!IsLoggedIn())
+            {
+                return RedirectToPage("/Account/Login");
+            }
             Infant infant = await context.Infants.FindAsync(id);
+            if (!IsInfantOwner(infant))
+            {
+                return RedirectToPage("/Error/Error404");
+            }
             InfantViewModel model = InfantViewModelFactory.Edit(infant);
             return View("InfantEditor", model);
         }
 
         [HttpPost]
-        public async Task<IActionResult> Edit ([FromForm]Infant infant)
+        public async Task<IActionResult> Edit (long id, [FromForm]Infant infant)
         {
+            if (!IsLoggedIn())
+            {
+                return RedirectToPage("/Account/Login");
+            }
+            Infant preSaveInfant = await context.Infants.AsNoTracking().FirstOrDefaultAsync(i => i.InfantId == id);
+            if (!IsInfantOwner(preSaveInfant))
+            {
+                return RedirectToPage("/Error/Error404");
+            }
             if (ModelState.IsValid)
             {
                 context.Infants.Update(infant);
@@ -73,14 +123,31 @@ namespace BabyTracker.Controllers
         // HTTP GET
         public async Task<IActionResult> Delete (long id)
         {
+            if (!IsLoggedIn())
+            {
+                return RedirectToPage("/Account/Login");
+            }
             Infant infant = await context.Infants.FindAsync(id);
+            if (!IsInfantOwner(infant))
+            {
+                return RedirectToPage("/Error/Error404");
+            }
             InfantViewModel model = InfantViewModelFactory.Delete(infant);
             return View("InfantEditor", model);
         }
 
         [HttpPost]
-        public async Task<IActionResult> Delete(Infant infant)
+        public async Task<IActionResult> Delete(long id, Infant infant)
         {
+            if (!IsLoggedIn())
+            {
+                return RedirectToPage("/Account/Login");
+            }
+            Infant preSaveInfant = await context.Infants.AsNoTracking().FirstOrDefaultAsync(i => i.InfantId == id);
+            if (!IsInfantOwner(preSaveInfant))
+            {
+                return RedirectToPage("/Error/Error404");
+            }
             context.Infants.Remove(infant);
             await context.SaveChangesAsync();
             return RedirectToAction("Index");
